requirements.txt
fastapi
uvicorn
gunicorn
diffusers
transformers
torch
pillow
